Interval estimate accuracy measures

Interval estimate accuracy measures

Usage

winkler_score(lower, upper, actual, level = 95, na.rm = TRUE, ...)

MSIS(
  lower,
  upper,
  actual,
  train,
  level = 95,
  period,
  d = period == 1,
  D = period > 1,
  na.rm = TRUE,
  ...
)

interval_measures

Format

An object of class list of length 2.

Arguments

lower

A vector of lower bound of interval forecasts from either the validation or test data.

upper

A vector of upper bound of interval forecasts from either the validation or test data.

actual

A vector of responses matching the forecasts.

level

The level of the forecast interval.

na.rm

If TRUE, remove the missing values before calculating the accuracy measure.

...

Additional arguments for each measure.

train

A vector of responses used to train the model.

period

The seasonal period of the data.

d

Should the response model include a first difference?

D

Should the response model include a seasonal difference?