Point estimate accuracy measures

Point estimate accuracy measures

Usage

ME(resid, na.rm = TRUE)

MAE(resid, na.rm = TRUE, ...)

MSE(resid, na.rm = TRUE, ...)

RMSE(resid, na.rm = TRUE, ...)

MPE(resid, actual, na.rm = TRUE, ...)

MAPE(resid, actual, na.rm = TRUE, ...)

MASE(
  resid,
  train,
  demean = FALSE,
  na.rm = TRUE,
  period,
  d = period == 1,
  D = period > 1,
  ...
)

RMSSE(
  resid,
  train,
  demean = FALSE,
  na.rm = TRUE,
  period,
  d = period == 1,
  D = period > 1,
  ...
)

point_measures

Format

An object of class list of length 8.

Arguments

resid

A vector of residuals from either the validation or test data.

na.rm

If TRUE, remove the missing values before calculating the accuracy measure.

...

Additional arguments for each measure.

actual

A vector of responses matching the forecasts.

train

A vector of responses used to train the model.

demean

Should the response be demeaned (for MASE and RMSSE)?

period

The seasonal period of the data.

d

Should the response model include a first difference?

D

Should the response model include a seasonal difference?